+++ src/components/table.jsx
@@ -1,11 +1,22 @@
 import React from 'react';
+import TableHeader from './tableHeader';
+import TableBody from './tableBody';
+import PropTypes from 'prop-types';
 
-const Table = () => {
+const Table = ({ onSort, selectedSort, columns, data }) => {
     return (
-        <div>
+        <table className='table'>
+            <TableHeader onSort={onSort} selectedSort={selectedSort} columns={columns}/>
+            <TableBody data={data} columns={columns}/>
+        </table>
 
-        </div>
     );
 };
 
+Table.propTypes = {
+    onSort: PropTypes.func,
+    selectedSort: PropTypes.object,
+    columns: PropTypes.object,
+    data: PropTypes.array
+};
 export default Table;
